Improve code quality of ensureLoggedIn()
