+++ frontend/src/Util.js
@@ -1,0 +1,11 @@
+import axios from "axios";
+import { URL_USER_SVC } from "./configs";
+
+export const ensureLoggedIn = async (navigate) => {
+  await axios.post(URL_USER_SVC + '/auth',
+      { withCredentials: true, credentials: 'include' })
+      .catch((err) => {
+          console.log(err)
+          navigate("/login")
+      });
+}
