+++ frontend/src/components/ChangePasswordPage.js
@@ -6,7 +6,7 @@ import { URL_USER_SVC } from "../configs";
 import { STATUS_CODE_BADREQUEST, STATUS_CODE_OK, STATUS_CODE_UNAUTHORIZED } from "../constants";
 import Cookies from 'js-cookie';
 import { useNavigate } from "react-router-dom";
-import { ensureLoggedIn } from '../customHooks';
+import { ensureLoggedIn } from '../Util';
 
 function ChangePasswordPage() {
     const [username, setUsername] = useState(Cookies.get('username'));
@@ -21,15 +21,6 @@ function ChangePasswordPage() {
     const [successMsg, setSuccessMsg] = useState("")
 
     let navigate = useNavigate();
-
-    // const ensureLoggedIn = async () => {
-    //     await axios.post(URL_USER_SVC + '/auth',
-    //         { withCredentials: true, credentials: 'include' })
-    //         .catch((err) => {
-    //             console.log(err)
-    //             navigate("/login")
-    //         });
-    // }
 
     useEffect(() => {
         ensureLoggedIn(navigate);
