+++ frontend/src/components/DeleteAccountPage.js
@@ -1,23 +1,13 @@
 import React, { useEffect } from 'react';
 import { Box, Typography } from "@mui/material";
-import axios from "axios";
-import { URL_USER_SVC } from "../configs";
 import { useNavigate } from "react-router-dom";
+import { ensureLoggedIn } from '../Util';
 
 function DeleteAccountPage(props) {
     let navigate = useNavigate();
 
-    const ensureLoggedIn = async () => {
-        await axios.post(URL_USER_SVC + '/auth',
-            { withCredentials: true, credentials: 'include' })
-            .catch((err) => {
-                console.log(err)
-                navigate("/login")
-            });
-    }
-
     useEffect(() => {
-        ensureLoggedIn();
+        ensureLoggedIn(navigate);
     })
 
     return (
