+++ frontend/src/components/MatchingService/RoomPage.js
@@ -6,25 +6,14 @@ import Grid from '@mui/material/Grid';
 import Stack from '@mui/material/Stack';
 import Typography from "@mui/material/Typography";
 import LogoutIcon from '@mui/icons-material/Logout';
-import axios from "axios";
-import { URL_USER_SVC } from "../../configs";
 import { useNavigate } from "react-router-dom";
-
+import { ensureLoggedIn } from '../../Util';
 
 export default function RoomPage() {
     let navigate = useNavigate();
 
-    const ensureLoggedIn = async () => {
-        await axios.post(URL_USER_SVC + '/auth',
-            { withCredentials: true, credentials: 'include' })
-            .catch((err) => {
-                console.log(err)
-                navigate("/login")
-            });
-    }
-
     useEffect(() => {
-        ensureLoggedIn();
+        ensureLoggedIn(navigate);
     })
 
     return (
